# chore(dataobj): Decode values in batches