--- a/pkg/dataobj/internal/dataset/value_encoding_plain_test.go
+++ b/pkg/dataobj/internal/dataset/value_encoding_plain_test.go
@@ -19,12 +19,15 @@ var testStrings = []string{
 	"baz",
 }
 
+var batchSize = 64
+
 func Test_plainEncoder(t *testing.T) {
 	var buf bytes.Buffer
 
 	var (
-		enc = newPlainEncoder(&buf)
-		dec = newPlainDecoder(&buf)
+		enc    = newPlainEncoder(&buf)
+		dec    = newPlainDecoder(&buf)
+		decBuf = make([]Value, batchSize)
 	)
 
 	for _, v := range testStrings {
@@ -34,13 +37,15 @@ func Test_plainEncoder(t *testing.T) {
 	var out []string
 
 	for {
-		str, err := dec.Decode()
+		n, err := dec.Decode(decBuf[:batchSize])
 		if errors.Is(err, io.EOF) {
 			break
 		} else if err != nil {
 			t.Fatal(err)
 		}
-		out = append(out, str.String())
+		for _, v := range decBuf[:n] {
+			out = append(out, v.String())
+		}
 	}
 
 	require.Equal(t, testStrings, out)
@@ -50,8 +55,9 @@ func Test_plainEncoder_partialRead(t *testing.T) {
 	var buf bytes.Buffer
 
 	var (
-		enc = newPlainEncoder(&buf)
-		dec = newPlainDecoder(&oneByteReader{&buf})
+		enc    = newPlainEncoder(&buf)
+		dec    = newPlainDecoder(&oneByteReader{&buf})
+		decBuf = make([]Value, batchSize)
 	)
 
 	for _, v := range testStrings {
@@ -61,13 +67,15 @@ func Test_plainEncoder_partialRead(t *testing.T) {
 	var out []string
 
 	for {
-		str, err := dec.Decode()
+		n, err := dec.Decode(decBuf[:batchSize])
 		if errors.Is(err, io.EOF) {
 			break
 		} else if err != nil {
 			t.Fatal(err)
 		}
-		out = append(out, str.String())
+		for _, v := range decBuf[:n] {
+			out = append(out, v.String())
+		}
 	}
 
 	require.Equal(t, testStrings, out)
@@ -87,19 +95,20 @@ func Benchmark_plainDecoder_Decode(b *testing.B) {
 	buf := bytes.NewBuffer(make([]byte, 0, 1024)) // Large enough to avoid reallocations.
 
 	var (
-		enc = newPlainEncoder(buf)
-		dec = newPlainDecoder(buf)
+		enc    = newPlainEncoder(buf)
+		dec    = newPlainDecoder(buf)
+		decBuf = make([]Value, batchSize)
 	)
 
 	for _, v := range testStrings {
 		require.NoError(b, enc.Encode(StringValue(v)))
 	}
 
+	var err error
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		for {
-			var err error
-			_, err = dec.Decode()
+			_, err = dec.Decode(decBuf[:batchSize])
 			if errors.Is(err, io.EOF) {
 				break
 			} else if err != nil {