**What this PR does / why we need it**:

**Which issue(s) this PR fixes**:

**Special notes for your reviewer**:

```console
$ benchstat bench_bitmapdecoder_old.txt bench_bitmapdecoder_new.txt 
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/v3/pkg/dataobj/internal/dataset
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
                                               │ bench_bitmapdecoder_old.txt │      bench_bitmapdecoder_new.txt      │
                                               │           sec/op            │    sec/op      vs base                │
_bitmapDecoder/width=1/variance=none-8                          1.674n ±  5%    1.557n ±  6%   -6.99% (p=0.001 n=10)
_bitmapDecoder/width=1/variance=alternating-8                   1.528n ±  2%    1.727n ±  6%  +13.02% (p=0.001 n=10)
_bitmapDecoder/width=1/variance=random-8                        1.722n ±  5%    1.560n ±  6%   -9.35% (p=0.000 n=10)
_bitmapDecoder/width=3/variance=none-8                          1.766n ± 15%    1.604n ±  5%   -9.20% (p=0.000 n=10)
_bitmapDecoder/width=3/variance=alternating-8                   9.068n ±  6%    9.236n ±  3%        ~ (p=0.631 n=10)
_bitmapDecoder/width=3/variance=random-8                        9.223n ±  6%   12.115n ± 28%        ~ (p=0.063 n=10)
_bitmapDecoder/width=5/variance=none-8                          1.801n ±  5%    1.506n ±  6%  -16.38% (p=0.000 n=10)
_bitmapDecoder/width=5/variance=alternating-8                   11.67n ±  6%    11.76n ±  3%        ~ (p=0.985 n=10)
_bitmapDecoder/width=5/variance=random-8                        12.36n ± 36%    12.94n ± 33%        ~ (p=0.645 n=10)
_bitmapDecoder/width=8/variance=none-8                          1.830n ± 41%    1.564n ±  7%  -14.51% (p=0.000 n=10)
_bitmapDecoder/width=8/variance=alternating-8                   11.44n ±  7%    11.63n ±  6%        ~ (p=0.470 n=10)
_bitmapDecoder/width=8/variance=random-8                        11.88n ± 46%    13.84n ± 31%        ~ (p=0.255 n=10)
_bitmapDecoder/width=32/variance=none-8                         2.544n ± 29%    1.518n ±  9%  -40.30% (p=0.000 n=10)
_bitmapDecoder/width=32/variance=alternating-8                  16.95n ±  5%    17.05n ± 86%        ~ (p=0.631 n=10)
_bitmapDecoder/width=32/variance=random-8                       17.37n ± 34%    19.29n ± 25%        ~ (p=0.225 n=10)
_bitmapDecoder/width=64/variance=none-8                         1.835n ±  3%    1.584n ± 43%        ~ (p=0.137 n=10)
_bitmapDecoder/width=64/variance=alternating-8                  21.66n ± 26%    19.30n ± 33%        ~ (p=0.143 n=10)
_bitmapDecoder/width=64/variance=random-8                       20.79n ± 32%    22.02n ± 21%        ~ (p=0.684 n=10)
geomean                                                         5.562n          5.387n         -3.15%
```

```console
$ benchstat bench_deltadecoder_old.txt bench_deltadecoder_new.txt
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/v3/pkg/dataobj/internal/dataset
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
                                     │ bench_deltadecoder_old.txt │     bench_deltadecoder_new.txt      │
                                     │           sec/op           │    sec/op     vs base               │
_deltaDecoder_Decode/Sequential-8                    6.743n ±  5%   6.645n ±  5%       ~ (p=0.404 n=10)
_deltaDecoder_Decode/Largest_delta-8                 29.61n ± 44%   29.07n ±  2%       ~ (p=0.183 n=10)
_deltaDecoder_Decode/Random-8                        32.15n ± 25%   31.26n ± 42%       ~ (p=0.796 n=10)
geomean                                              18.58n         18.21n        -2.01%
```

```console
$ benchstat bench_plaindecoder_old.txt bench_plaindecoder_new.txt
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/v3/pkg/dataobj/internal/dataset
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
                       │ bench_plaindecoder_old.txt │     bench_plaindecoder_new.txt      │
                       │           sec/op           │   sec/op     vs base                │
_plainDecoder_Decode-8                  10.38n ± 5%   18.62n ± 4%  +79.47% (p=0.000 n=10)
```